+++ src/renderer/i18n/en/common.ts
@@ -96,6 +96,7 @@ const common: CommonMessages = {
   'common.min': 'Min',
   'common.search': 'Search',
   'common.searchAsset': 'Search Asset',
+  'common.addAssetManually': 'ERC20 not showing? add asset manually',
   'common.searchExample': 'Search example for non synth chain.ticker i.e btc.btc or for synth btc/btc',
   'common.excludeSynth': 'Exclude Synths',
   'common.retry': 'Retry',