+++ src/renderer/i18n/ru/common.ts
@@ -96,6 +96,7 @@ const common: CommonMessages = {
   'common.min': 'Мин.',
   'common.search': 'Поиск',
   'common.searchAsset': 'Поиск актива',
+  'common.addAssetManually': 'ERC20 не отображается? Добавьте актив вручную',
   'common.searchExample': 'Пример поиска для не синтетической цепи.ticker, например, btc.btc или для синтетики btc/btc',
   'common.excludeSynth': 'Исключить синтезы',
   'common.retry': 'Повторить',