+++ CHANGELOG.md
@@ -4,7 +4,9 @@
 
 ## Add
 
-- Chain enabled disable [#281](https://github.com/asgardex/asgardex-desktop/pull/309)
+- Chain enabled disable [#281](https://github.com/asgardex/asgardex-desktop/pull/312)
+- Added erc20 manual add [#313] (https://github.com/asgardex/asgardex-desktop/pull/314)
+
 # 1.22.0 (2024-07-29)
 
 ## Updates/Fixes